requirements.txt
# Core Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
alembic>=1.13.0
google-adk>=0.5.0
google-generativeai>=0.8.0
chromadb>=0.5.0
structlog>=24.0.0
tenacity>=8.3.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0

# Development Dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.5.0
mypy>=1.10.0
